requirements.txt
pyrogram
pyTelegramBotAPI
python-dotenv